Code cleanup : - Removed "final" for static and private methods; - for (int i = 0; i < ar.length; i++) replaced with "for each" where possible - removed unnecessary type/class castings; - a.toArray(new T[0]) replaced with a.toArray(new T[a.size()]);

+++ src/main/java/org/jcodec/codecs/aac/BlockReader.java
@@ -21,7 +21,7 @@ public class BlockReader {
         if (type == TYPE_END)
             return null;
 
-        int id = (int) bits.readNBit(4);
+        int id = bits.readNBit(4);
 
         return null;
     }
