+++ src/main/java/org/jcodec/codecs/aac/blocks/BlockDSE.java
@@ -14,9 +14,9 @@ import org.jcodec.common.io.BitReader;
 public class BlockDSE extends Block {
 
     public void parse(BitReader in) {
-        int elemType = (int) in.readNBit(4);
+        int elemType = in.readNBit(4);
         int byte_align = in.read1Bit();
-        int count = (int) in.readNBit(8);
+        int count = in.readNBit(8);
         if (count == 255)
             count += in.readNBit(8);
         if (byte_align != 0)
