+++ src/main/java/org/jcodec/movtool/streaming/tracks/MTSTrackFactory.java
@@ -73,7 +73,7 @@ public class MTSTrackFactory {
                     ByteBuffer tsBuf = NIOUtils.read(buf, 188);
                     Assert.assertEquals(0x47, tsBuf.get() & 0xff);
                     int guidFlags = ((tsBuf.get() & 0xff) << 8) | (tsBuf.get() & 0xff);
-                    int guid = (int) guidFlags & 0x1fff;
+                    int guid = guidFlags & 0x1fff;
                     if (guid == targetGuid) {
                         int b0 = tsBuf.get() & 0xff;
                         int counter = b0 & 0xf;
